Node Page Fixes & Minor Fixes
Fixed an issue with the node page where it didnt support the data in the peers table.
Modified the node info model to support the peers table.
Modified the generate nodes info function to query the peers_v2 table.
Removed async from routes that were using the synchronous Cassandra driver.